Separación de funcionalidades DAO y Controller
