+++ src/backend/ServicesDeskUCABWS/BussinesLogic/DAO/DepartamentoDAO/IDepartamentoDAO.cs
@@ -15,11 +15,7 @@ namespace ServicesDeskUCABWS.BussinesLogic.DAO.DepartamentoDAO
         public DepartamentoDto ConsultarPorID(Guid id);
         public DepartamentoDto eliminarDepartamento(Guid id);
         public DepartamentoDto_Update ActualizarDepartamento(Departamento departamento);
-        public List<DepartamentoDto> GetByIdDepartamento(Guid idGrupo);
-        public List<string> AsignarGrupoToDepartamento(Guid id, string idDept);
         public List<DepartamentoDto> DeletedDepartamento();
         public List<DepartamentoDto> NoAsociado();
-        public List<string> EditarRelacion(Guid id, string idDepartamentos);
-
     }
 }
