+++ src/backend/ServicesDeskUCABWS/BussinesLogic/DAO/GrupoDAO/IGrupoDAO.cs
@@ -1,4 +1,5 @@
 ﻿using ServicesDeskUCABWS.BussinesLogic.DTO.GrupoDTO;
+using ServicesDeskUCABWS.BussinesLogic.DTO.DepartamentoDTO;
 using ServicesDeskUCABWS.Entities;
 using System;
 using System.Collections.Generic;
@@ -15,5 +16,8 @@ namespace ServicesDeskUCABWS.BussinesLogic.DAO.GrupoDAO
         public GrupoDto_Update ModificarGrupoDao(Grupo grupo);
         public bool QuitarAsociacion(Guid grupoId);
         public List<GrupoDto> ConsultarGrupoNoEliminado();
+        public List<string> AsignarGrupoToDepartamento(Guid id, string idDept);
+        public List<string> EditarRelacion(Guid id, string idDepartamentos);
+        public List<DepartamentoDto> GetByIdDepartamento(Guid idGrupo);
     }
 }
