+++ src/backend/ServicesDeskUCABWS/Controllers/ControllerDepartamento/DepartamentoController.cs
@@ -146,24 +146,6 @@ namespace ServicesDeskUCABWS.Controllers.ControllerDepartamento
             return response;
         }
 
-        [HttpPut]
-        [Route("AsignarGrupoToDepartamento/{id}")]
-        public ApplicationResponse<List<string>> AsignarGrupoToDepartamento([FromRoute] Guid id, [FromBody] string idDepartamentos)
-        {
-            var response = new ApplicationResponse<List<string>>();
-            try
-            {
-                response.Data = _departamentoDAO.AsignarGrupoToDepartamento(id, idDepartamentos);
-            }
-            catch (ExceptionsControl ex)
-            {
-                response.Success = false;
-                response.Message = ex.Mensaje;
-                response.Exception = ex.Excepcion.ToString();
-            }
-            return response;
-        }
-
         [HttpGet("ConsultarDepartamentoNoAsociado/")]
         public ApplicationResponse<List<DepartamentoDto>> ListaDepartamentoNoAsociado()
         {
@@ -190,24 +172,6 @@ namespace ServicesDeskUCABWS.Controllers.ControllerDepartamento
             try
             {
                 response.Data = _departamentoDAO.DeletedDepartamento();
-            }
-            catch (ExceptionsControl ex)
-            {
-                response.Success = false;
-                response.Message = ex.Mensaje;
-                response.Exception = ex.Excepcion.ToString();
-            }
-            return response;
-        }
-
-        [HttpPut]
-        [Route("EditarRelacion/{id}")]
-        public ApplicationResponse<List<string>> EditarRelacion([FromRoute] Guid id, [FromBody] string idDepartamentos)
-        {
-            var response = new ApplicationResponse<List<string>>();
-            try
-            {
-                response.Data = _departamentoDAO.EditarRelacion(id, idDepartamentos);
             }
             catch (ExceptionsControl ex)
             {
