+++ src/backend/ServicesDeskUCABWS/Controllers/ControllerGrupo/GrupoController.cs
@@ -149,5 +149,58 @@ namespace ServicesDeskUCABWS.Controllers.ControllerGrupo
 			return response;
 		}
 
-	}
+        [HttpPut]
+        [Route("AsignarGrupoToDepartamento/{id}")]
+        public ApplicationResponse<List<string>> AsignarGrupoToDepartamento([FromRoute] Guid id, [FromBody] string idDepartamentos)
+        {
+            var response = new ApplicationResponse<List<string>>();
+            try
+            {
+                response.Data = _grupoDAO.AsignarGrupoToDepartamento(id, idDepartamentos);
+            }
+            catch (ExceptionsControl ex)
+            {
+                response.Success = false;
+                response.Message = ex.Mensaje;
+                response.Exception = ex.Excepcion.ToString();
+            }
+            return response;
+        }
+
+        [HttpPut]
+        [Route("EditarRelacion/{id}")]
+        public ApplicationResponse<List<string>> EditarRelacion([FromRoute] Guid id, [FromBody] string idDepartamentos)
+        {
+            var response = new ApplicationResponse<List<string>>();
+            try
+            {
+                response.Data = _grupoDAO.EditarRelacion(id, idDepartamentos);
+            }
+            catch (ExceptionsControl ex)
+            {
+                response.Success = false;
+                response.Message = ex.Mensaje;
+                response.Exception = ex.Excepcion.ToString();
+            }
+            return response;
+        }
+
+        [HttpGet("ConsultarDepartamentosPorIdGrupo/{idGrupo}")]
+        public ApplicationResponse<List<DepartamentoDto>> ListaDepartamentosGrupo(Guid idGrupo)
+        {
+            var response = new ApplicationResponse<List<DepartamentoDto>>();
+            try
+            {
+                response.Data = _grupoDAO.GetByIdDepartamento(idGrupo);
+            }
+            catch (ExceptionsControl ex)
+            {
+                response.Success = false;
+                response.Message = ex.Mensaje;
+                response.Exception = ex.Excepcion.ToString();
+            }
+            return response;
+        }
+
+    }
 }
