+++ src/frontend/ServiceDeskUCAB/Views/Grupo/Index.cshtml
@@ -70,7 +70,6 @@
             })
         })
     </script>
-
 }
 
     <!-- Placeholders de las ventanas emergentes (Grupo)-->
